ugh, got stuck of ssbos. Solution below.
SO the reason I was confused was that the ssbo wasnt getting written
to. It wasnt getting written to as there werent any fragments for the
fragment shader to run. There werent fragments as the verts were
culled. The verts were culled as they were outside the viewport.

So dumb. I need to read up on clipping and also fix some glsl-spec
issues we would have hit if I hadnt got stuck :)